Removed cyclic imports
Removes the import cycle of index.js importing withSafeArea which imports
index.js, and so on.
